Refactor import script
Move helper functions into jh_cleaning and JHCrud
